refactor: Add an authmiddleware for user garden and virtual garden controllers

+++ app/controllers/userGarden.controller.js
@@ -3,7 +3,7 @@ import datamapper from '../datamappers/userGarden.datamapper.js';
 const controller = {
   getFavorites: async (req, res) => {
     try {
-      const userId = req.query.id;
+      const userId = req.userId;
       const data = await datamapper.getAllFavorites(userId);
       if (data.length === 0) {
         return res.status(200).json('No favories found.');
@@ -16,7 +16,7 @@ const controller = {
 
   addFavorite: async (req, res) => {
     try {
-      const userId = req.body.user_id;
+      const userId = req.userId;
       const productToAdd = req.body.product_id;
       const data = await datamapper.addFavorite(productToAdd, userId);
       res.json(data);
@@ -33,8 +33,8 @@ const controller = {
       if (!productToRemove) {
         return res.status(400).json({ message: "Product to remove is required" });
       }
-      if (!userId) {
-        return res.status(400).json({ message: "User is required" });
+      if (!userId || userId !== req.userId) {
+        return res.status(400).json({ message: "Unauthorized user" });
       }
 
       await datamapper.removeFavorite(userId, productToRemove);
